\name{numerical_experiments_data}
\title{Data generated by the demo 'numerical_experiments' of the package 'nvmix'}
\alias{numerical_experiments_data}
\docType{data}
\usage{data(numerical_experiments_data, package = "nvmix")}
\description{
  Data generated by the demo \code{numerical_experiments} of the \code{nvmix} package.
}
\format{
  A list with 10 elements:
  \describe{
    \item{\code{$pnvmix.abserrors}}{Array as output by \code{pnvmix_testing_abserr()},
           see Sec. 1.1 of \code{demo(numerical_experiments)}.}  
    \item{\code{$pnvmix.t.variances}}{Array as output by \code{precond_testing_variance()},
           see Sec. 1.1 of \code{demo(numerical_experiments)}.}
    \item{\code{$pnvmix.t.sobolind}}{Array as output by \code{pnvmix_estimate_sobolind()},
           see Sec. 1.1 of \code{demo(numerical_experiments)}.}
    \item{\code{$pnvmix.t.timing}}{Array as output by \code{pnvmix_timing_mvt()},
           see Sec. 1.1 of \code{demo(numerical_experiments)}.}
    \item{\code{$dnvmix.results}}{Array as output by \code{dnvmix_testing()},
           see Sec. 1.2 of \code{demo(numerical_experiments)}.}
    \item{\code{$fitnvmix.results}}{Array as output by \code{fitnvmix_testing()},
           see Sec. 1.3 of \code{demo(numerical_experiments)}.}
    \item{\code{$fit.dj30.anaylytical}}{Array containing results of \code{fitnvmix()}
          applied to DJ30 data using analytical weights/densities, see Sec. 5 of 
          \code{demo(numerical_experiments)}.}  
    \item{\code{$fit.dj30.estimated}}{Array containing results of \code{fitnvmix()}
          applied to DJ30 data using estimated weights/densities, see Sec. 5 of 
          \code{demo(numerical_experiments)}.}  
    \item{\code{$qqplots.dj30}}{Array containing results of \code{qqplot.maha()}
          applied to DJ30 data, see Sec. 5 of \code{demo(numerical_experiments)}.}  
    \item{\code{$tailprobs.dj30}}{Array containing estimated quantile shortfall
          probabilities of models fitted to DJ30 data, see Sec. 5 of 
          \code{demo(numerical_experiments)}.}
  }
}
\references{
  Hintz, E., Hofert, M. and Lemieux, C. (2019),
  Normal variance mixtures: Distribution, density and parameter estimation.
  \url{https://arxiv.org/abs/1911.03017}. 
}
\keyword{datasets}