\name{numerical_experiments_data}
\title{Data Generated by the Demo 'numerical_experiments'}
\alias{numerical_experiments_data}
\docType{data}
\usage{data(numerical_experiments_data, package = "nvmix")}
\description{
  Data generated by the \code{demo('numerical_experiments')} of the
  \code{nvmix} package.
}
\format{
  A list with 10 elements:
  \describe{
    \item{\code{$pnvmix.abserrors}}{Array as returned by the function
      \code{pnvmix_testing_abserr()}, 
      see Section 1.1 of the \code{demo('numerical_experiments')}.}
    \item{\code{$pnvmix.t.variances}}{Array as returned by the function
      \code{precond_testing_variance()},
      see Section 1.1 of the \code{demo('numerical_experiments')}.}
    \item{\code{$pnvmix.t.sobolind}}{Array as returned by the function
      \code{pnvmix_estimate_sobolind()},
      see Section 1.1 of the \code{demo('numerical_experiments')}.}
    \item{\code{$pnvmix.t.timing}}{Array as returned by the function
      \code{pnvmix_timing_mvt()},
      see Section 1.1 of the \code{demo('numerical_experiments')}.}
    \item{\code{$dnvmix.results}}{Array as returned by the function
      \code{dnvmix_testing()},
      see Section 1.2 of the \code{demo('numerical_experiments')}.}
    \item{\code{$fitnvmix.results}}{Array as returned by the function
      \code{fitnvmix_testing()},
      see Section 1.3 of the \code{demo('numerical_experiments')}.}
    \item{\code{$fit.dj30.anaylytical}}{Array containing results of
      \code{fitnvmix()} applied to DJ30 data using analytical
      weights/densities, see Section 5 of
      \code{demo('numerical_experiments')}.}
    \item{\code{$fit.dj30.estimated}}{Array containing results of
      \code{fitnvmix()} applied to DJ30 data using estimated
      weights/densities, see Section 5 of
      \code{demo('numerical_experiments')}.}
    \item{\code{$qqplots.dj30}}{Array containing results of \code{qqplot.maha()}
      applied to DJ30 data, see Section 5 of the \code{demo('numerical_experiments')}.}
    \item{\code{$tailprobs.dj30}}{Array containing estimated quantile shortfall
      probabilities of models fitted to DJ30 data, see Section 5 of
      \code{demo('numerical_experiments')}.}
  }
}
\references{
  Hintz, E., Hofert, M. and Lemieux, C. (2021),
  Normal variance mixtures: Distribution, density and parameter estimation.
  \emph{Computational Statistics and Data Analysis} 157C, 107175.
}
\keyword{datasets}